Implementation exercises

+++ app/src/main/java/com/example/samurairoad/room/tables/WorkoutDataTableModel.kt
@@ -5,6 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 
+//TODO change cascade on ex
 @Entity(tableName = "workout_data", foreignKeys = [
     ForeignKey(
         entity = WorkoutTableModel::class,
@@ -22,15 +23,15 @@ import androidx.room.PrimaryKey
 data class WorkoutDataTableModel(
 
     @ColumnInfo(name = "workoutId")
-    var WorkoutID: Int,
+    var WorkoutID: Long,
 
     @ColumnInfo(name = "exerciseId")
-    var ExerciseID: Int,
+    var ExerciseID: Long,
 
 ){
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id")
-    var Id: Int? = null
+    var Id: Long? = null
 
 }
